requirements.txt
opencv-python>=4.5.5
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.1.0
plyer>=2.0.0
PyQt5>=5.15.0
